Fix inline diagnostics in the project diff (#26275)
https://github.com/zed-industries/zed/blob/63465bd52c340c218c0b403d48e83ec69c21f694/crates/editor/src/element.rs#L1643

Due to the snippet above, Zed is supposed to have `row` larger or equal
to `start_row` here:


https://github.com/zed-industries/zed/blob/63465bd52c340c218c0b403d48e83ec69c21f694/crates/editor/src/element.rs#L1694

yet the panic were reported when clicking in the project diff.

That project diff has a lot of highlighting happening already, so the PR
disables inline diagnostics within a git diff view.


Release Notes:

- N/A